Move some S3 code to shared module
