+++ python/lsst/resources/s3.py
@@ -21,18 +21,24 @@ import threading
 
 __all__ = ("S3ResourcePath",)
 
-from http.client import HTTPException, ImproperConnectionState
-from types import ModuleType
-from typing import IO, TYPE_CHECKING, Any, Callable, Iterator, List, Optional, Tuple, Union, cast
+from typing import IO, TYPE_CHECKING, Iterator, List, Optional, Tuple, Union
 
 from botocore.exceptions import ClientError
 from lsst.utils.timer import time_this
-from urllib3.exceptions import HTTPError, RequestError
 
 from ._resourceHandles._baseResourceHandle import ResourceHandleProtocol
 from ._resourceHandles._s3ResourceHandle import S3ResourceHandle
 from ._resourcePath import ResourcePath
-from .s3utils import bucketExists, getS3Client, s3CheckFileExists
+from .s3utils import (
+    bucketExists,
+    getS3Client,
+    s3CheckFileExists,
+    backoff,
+    retryable_io_errors,
+    all_retryable_errors,
+    max_retry_time,
+    _TooManyRequestsException
+)
 
 if TYPE_CHECKING:
     try:
@@ -40,64 +46,6 @@ if TYPE_CHECKING:
     except ImportError:
         pass
     from .utils import TransactionProtocol
-
-# https://pypi.org/project/backoff/
-try:
-    import backoff
-except ImportError:
-
-    class Backoff:
-        @staticmethod
-        def expo(func: Callable, *args: Any, **kwargs: Any) -> Callable:
-            return func
-
-        @staticmethod
-        def on_exception(func: Callable, *args: Any, **kwargs: Any) -> Callable:
-            return func
-
-    backoff = cast(ModuleType, Backoff)
-
-
-class _TooManyRequestsException(Exception):
-    """Private exception that can be used for 429 retry.
-
-    botocore refuses to deal with 429 error itself so issues a generic
-    ClientError.
-    """
-
-    pass
-
-
-# settings for "backoff" retry decorators. these retries are belt-and-
-# suspenders along with the retries built into Boto3, to account for
-# semantic differences in errors between S3-like providers.
-retryable_io_errors = (
-    # http.client
-    ImproperConnectionState,
-    HTTPException,
-    # urllib3.exceptions
-    RequestError,
-    HTTPError,
-    # built-ins
-    TimeoutError,
-    ConnectionError,
-    # private
-    _TooManyRequestsException,
-)
-
-# Client error can include NoSuchKey so retry may not be the right
-# thing. This may require more consideration if it is to be used.
-retryable_client_errors = (
-    # botocore.exceptions
-    ClientError,
-    # built-ins
-    PermissionError,
-)
-
-# Combine all errors into an easy package. For now client errors
-# are not included.
-all_retryable_errors = retryable_io_errors
-max_retry_time = 60
 
 
 log = logging.getLogger(__name__)
